Issue #2747911 Convert UninstallTest to BrowserTestBase
Move UninstallTest to its proper place
